Handle disks that have partitions and labels
If an error occurs while installing OS, or if there are existing
partitions the set-partition-script will now aggressively reset
the disk to a blank state.

If the device is not mounted in the install container, or any running
system containers, the script will blow away the MBR and partition table
and configure a single primary bootable partition.